Complete rework of the HUD base and HUD frames
Includes the original FreeCAD model as well as STL files for a single part base and single part left and right frames, as well as sliced (split) parts for base and frame for those that have smaller 3D printers (like me).  Removed old files that used the old design.  The new design is more accurate to the real thing.