Reworked left and right consoles.
Reworked left and right consoles.  Note that this is still a work in progress.  I need to design 3D printed  brackets for between the upper frame members and for providing mounting solutions to the panels and what not.  I also need to redesign the canopy ring, which will make it more accurate but might possibly require a backing piece to be cut from steel for strength.